+++ core/src/test/java/arez/test/ComputedValueApiTest.java
@@ -5,7 +5,9 @@ import arez.Arez;
 import arez.ArezContext;
 import arez.ComputedValue;
 import arez.Flags;
+import arez.SafeFunction;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicReference;
 import org.testng.annotations.Test;
 import static org.testng.Assert.*;
 
@@ -26,7 +28,7 @@ public class ComputedValueApiTest
     final ComputedValue<Integer> computedValue = context.computed( () -> {
       computedCallCount.incrementAndGet();
       return result.get();
-    }, Flags.NON_AREZ_DEPENDENCIES );
+    }, Flags.AREZ_OR_EXTERNAL_DEPENDENCIES );
 
     assertEquals( computedCallCount.get(), 0 );
     assertEquals( observerCallCount.get(), 0 );
@@ -56,5 +58,79 @@ public class ComputedValueApiTest
     assertEquals( observerCallCount.get(), 2 );
 
     context.safeAction( () -> assertEquals( computedValue.get(), (Integer) 21 ) );
+  }
+
+  @Test
+  public void computedWithNoDependencies()
+    throws Exception
+  {
+    final AtomicInteger observerCallCount = new AtomicInteger();
+    final AtomicInteger computedCallCount = new AtomicInteger();
+
+    final ArezContext context = Arez.context();
+    final ComputedValue<Integer> computedValue = context.computed( () -> {
+      computedCallCount.incrementAndGet();
+      return 1;
+    }, Flags.AREZ_OR_NO_DEPENDENCIES );
+
+    assertEquals( computedCallCount.get(), 0 );
+    assertEquals( observerCallCount.get(), 0 );
+
+    context.observer( () -> {
+      observerCallCount.incrementAndGet();
+      computedValue.get();
+    } );
+
+    assertEquals( computedCallCount.get(), 1 );
+    assertEquals( observerCallCount.get(), 1 );
+
+    context.safeAction( () -> assertEquals( computedValue.get(), (Integer) 1 ) );
+  }
+
+  @Test
+  public void arezOrExternalDependencies()
+    throws Throwable
+  {
+    final ArezContext context = Arez.context();
+
+    final AtomicInteger calls = new AtomicInteger();
+    final AtomicReference<String> result = new AtomicReference<>();
+    result.set( "" );
+    final SafeFunction<String> action = () -> {
+      calls.incrementAndGet();
+      return result.get();
+    };
+    final ComputedValue<String> computedValue =
+      context.computed( "TestComputed", action, Flags.AREZ_OR_EXTERNAL_DEPENDENCIES | Flags.KEEPALIVE );
+
+    final AtomicInteger observerCallCount = new AtomicInteger();
+    context.observer( () -> {
+      observerCallCount.incrementAndGet();
+      computedValue.get();
+    } );
+
+    assertEquals( observerCallCount.get(), 1 );
+    assertEquals( calls.get(), 1 );
+
+    context.action( () -> assertEquals( computedValue.get(), "" ) );
+
+    assertEquals( observerCallCount.get(), 1 );
+    assertEquals( calls.get(), 1 );
+
+    context.action( computedValue::reportPossiblyChanged );
+
+    context.action( () -> assertEquals( computedValue.get(), "" ) );
+
+    assertEquals( observerCallCount.get(), 1 );
+    assertEquals( calls.get(), 2 );
+
+    result.set( "NewValue" );
+
+    context.action( computedValue::reportPossiblyChanged );
+
+    context.action( () -> assertEquals( computedValue.get(), "NewValue" ) );
+
+    assertEquals( observerCallCount.get(), 2 );
+    assertEquals( calls.get(), 3 );
   }
 }