+++ core/src/test/java/arez/test/ObserverApiTest.java
@@ -41,6 +41,17 @@ public class ObserverApiTest
   }
 
   @Test
+  public void autorun_noDependencies()
+    throws Exception
+  {
+    final AtomicInteger callCount = new AtomicInteger();
+
+    Arez.context().observer( callCount::incrementAndGet, Flags.AREZ_OR_NO_DEPENDENCIES );
+
+    assertEquals( callCount.get(), 1 );
+  }
+
+  @Test
   public void schedule_autorun_doesNotExecuteIfNotStale()
     throws Exception
   {
@@ -74,7 +85,7 @@ public class ObserverApiTest
     final Observer observer = context.observer( () -> {
       observable.reportObserved();
       callCount.incrementAndGet();
-    }, Flags.NON_AREZ_DEPENDENCIES );
+    }, Flags.AREZ_OR_EXTERNAL_DEPENDENCIES );
 
     assertEquals( callCount.get(), 1 );
 
@@ -95,7 +106,7 @@ public class ObserverApiTest
     final Observer observer = context.observer( () -> {
       observable.reportObserved();
       callCount.incrementAndGet();
-    }, onDepsChangedCallCount::incrementAndGet, Flags.NON_AREZ_DEPENDENCIES );
+    }, onDepsChangedCallCount::incrementAndGet, Flags.AREZ_OR_EXTERNAL_DEPENDENCIES );
 
     assertEquals( callCount.get(), 1 );
     assertEquals( onDepsChangedCallCount.get(), 0 );