+++ integration-tests/src/test/java/arez/integration/observed/ObservedArezDependenciesOnlyTest.java
@@ -44,7 +44,7 @@ public class ObservedArezDependenciesOnlyTest
 
     // reportStale should result in exception
     assertInvariantFailure( () -> safeAction( () -> component.getRenderObserver().reportStale() ),
-                            "Arez-0199: Observer.reportStale() invoked on observer named 'TestComponent1.0.render' but arezOnlyDependencies = true." );
+                            "Arez-0199: Observer.reportStale() invoked on observer named 'TestComponent1.0.render' but the observer has not specified AREZ_OR_EXTERNAL_DEPENDENCIES flag." );
 
     assertEquals( component._renderCallCount, 1 );
   }