+++ processor/src/test/resources/input/com/example/observed/ArezOrNoneDependenciesModel.java
@@ -5,10 +5,10 @@ import arez.annotations.DepType;
 import arez.annotations.Observed;
 
 @ArezComponent
-public abstract class NonArezDependenciesButNoObserverRefModel
+public abstract class ArezOrNoneDependenciesModel
 {
-  @Observed( depType = DepType.AREZ_OR_EXTERNAL )
-  void doStuff()
+  @Observed( depType = DepType.AREZ_OR_NONE )
+  void render()
   {
   }
 }